Make VacationMode.Model.State a union type.
This makes sure that a start time and end time can't be set
when the vacation mode is disabled.

The change also made a bug visible where JSON from the
server was not properly decoded when vacation mode was off,
because no start_time or end_time were rendered.

This commit also fixes that.